move AlertModal to folder Modal and rename
- it is not only an AlertModal, but Modal for other things too
- better to rename it as Modal
