+++ src/Screens/Main/Chat/chatProperties.ts
@@ -4,27 +4,19 @@ import { DropDownItem } from 'src/Screens/components/DropDownMenu';
 export const dialogProperties = {
   Delete: {
     messageId: 'main.chat.delete.confirmation',
-    primaryButtonMessage: 'main.chat.delete',
-    secondaryButtonMessage: 'meta.back',
     modalType: 'danger',
   },
   Unban: {
     messageId: 'main.chat.unban.confirmation',
     modalType: 'info',
-    primaryButtonMessage: 'main.chat.unban',
-    secondaryButtonMessage: 'meta.back',
   },
   Ban: {
     messageId: 'main.chat.ban.confirmation',
     modalType: 'warning',
-    primaryButtonMessage: 'main.chat.ban',
-    secondaryButtonMessage: 'meta.back',
   },
   Archive: {
     messageId: 'main.chat.archive.confirmation',
     modalType: 'info',
-    primaryButtonMessage: 'main.chat.archive',
-    secondaryButtonMessage: 'meta.back',
   },
 } as const;
