+++ src/Screens/components/Toast.tsx
@@ -9,7 +9,7 @@ import fonts from './fonts';
 import colors from './colors';
 import shadow from './shadow';
 
-import { props as toastProps } from './alertProps';
+import { props as toastProps } from './Modal/modalProps';
 
 export type AlertVariant = 'danger' | 'success' | 'info' | 'warning';
 type Props = {
